refactor(home): move herosection to app layer
