+++ service/app/src/app/_components/HeroSection.tsx
